v4: moved CreateTempDir to tesutils/helpers.go (for backwards comp.)
